chore: refactor server serialization code
