+++ apps/server/src/package.ts
@@ -4,4 +4,5 @@ export type * from "./modules/world/schema";
 export type * from "./context";
 
 // Except for the transformers which need to be used in both server and client at runtime
-export * from "./serialization";
+export * from "./serialization/selected";
+export * from "./tokenHeaderName";
