+++ apps/server/src/serialization/cbor.ts
@@ -2,7 +2,7 @@ import type { AbstractTransformer } from "@mp/transformer";
 import { cborTransformer, type ConcreteTransformer } from "@mp/transformer";
 import type { SocketIO_DTO } from "@mp/network/server";
 import { Vector } from "@mp/math";
-import type { ClientStateUpdate } from "./context";
+import type { ClientStateUpdate } from "../context";
 
 cborTransformer.registerTag(9001, Vector, {
   encode: ({ x, y }) => [x, y],
@@ -14,12 +14,10 @@ cborTransformer.registerTag(9002, Map, {
   decode: (entries) => new Map(entries),
 });
 
-export const serialization = {
+export const cborSerialization = {
   stateUpdate: cborTransformer as unknown as ConcreteTransformer<
     ClientStateUpdate,
     SocketIO_DTO
   >,
   rpc: cborTransformer as unknown as AbstractTransformer<SocketIO_DTO>,
 };
-
-export const tokenHeaderName = "token";
