+++ apps/server/src/serialization/json.ts
@@ -2,7 +2,7 @@ import type { AbstractTransformer } from "@mp/transformer";
 import { jsonTransformer, type ConcreteTransformer } from "@mp/transformer";
 import { Vector } from "@mp/math";
 import type { SocketIO_DTO } from "@mp/network/server";
-import type { ClientStateUpdate } from "./context";
+import type { ClientStateUpdate } from "../context";
 
 jsonTransformer.registerClass(Vector, { identifier: "vector" });
