+++ apps/server/src/serialization/selected.ts
@@ -1,0 +1,3 @@
+// This module is an abstraction of the currently selected serialization implementation.
+
+export { cborSerialization as serialization } from "./cbor";
