+++ apps/server/src/tokenHeaderName.ts
@@ -1,0 +1,1 @@
+export const tokenHeaderName = "token";
